Addressed PR review
